Delete useApi; Add useVisualChangeset and useAiCopySuggestion
