+++ src/visual_editor/components/AICopySuggestor.tsx
@@ -3,7 +3,7 @@ import React, { FC, MouseEvent, useCallback, useEffect, useState } from "react";
 import TextareaAutosize from "react-textarea-autosize";
 import { BiLoaderCircle } from "react-icons/bi";
 import { CopyMode } from "../../../devtools";
-import { TransformCopyFn } from "../lib/hooks/useApi";
+import { TransformCopyFn } from "../lib/hooks/useAiCopySuggestion";
 
 const AICopySuggestor: FC<{
   parentElement: Element;
