+++ src/visual_editor/lib/hooks/useVisualChangeset.ts
@@ -1,0 +1,154 @@
+import { useCallback, useEffect, useState } from "react";
+import {
+  ErrorCode,
+  CSPError,
+  VisualEditorVariation,
+  Message,
+  APIVisualChangeset,
+  APIExperiment,
+  LoadVisualChangesetRequestMessage,
+  UpdateVisualChangesetRequestMessage,
+} from "../../../../devtools";
+import normalizeVariations from "../normalizeVariations";
+type UseVisualChangesetHook = (visualChangesetId: string) => {
+  loading: boolean;
+  experimentUrl: string | null;
+  variations: VisualEditorVariation[];
+  updateVariationAtIndex: (
+    index: number,
+    updates: Partial<VisualEditorVariation>
+  ) => void;
+  error: ErrorCode | null;
+  cspError: CSPError | null;
+};
+
+const useVisualChangeset: UseVisualChangesetHook = (visualChangesetId) => {
+  const [error, setError] = useState<ErrorCode | null>(null);
+  const [loading, setLoading] = useState(false);
+  const [cspError, setCSPError] = useState<CSPError | null>(null);
+  const [visualChangeset, setVisualChangeset] =
+    useState<APIVisualChangeset | null>(null);
+  const [experiment, setExperiment] = useState<APIExperiment | null>(null);
+  const [experimentUrl, setExperimentUrl] = useState<string | null>(null);
+  const [variations, setVariations] = useState<VisualEditorVariation[]>([]);
+
+  // TODO refine this to only catch inline script errors
+  document.addEventListener("securitypolicyviolation", (e) => {
+    setError("csp-error");
+    setCSPError({
+      violatedDirective: e.violatedDirective,
+    });
+  });
+
+  const updateVisualChangeset = useCallback(
+    async (variations: VisualEditorVariation[]) => {
+      setLoading(true);
+      setError(null);
+
+      const updatePayload: Partial<APIVisualChangeset> = {
+        visualChanges: variations.map((v) => ({
+          variation: v.variationId,
+          domMutations: v.domMutations,
+          css: v.css,
+          js: v.js,
+          description: v.description,
+        })),
+      };
+
+      const updateVisualChangesetMessage: UpdateVisualChangesetRequestMessage =
+        {
+          type: "GB_REQUEST_UPDATE_VISUAL_CHANGESET",
+          data: {
+            visualChangesetId,
+            updatePayload,
+          },
+        };
+
+      window.postMessage(updateVisualChangesetMessage, window.location.origin);
+    },
+    [visualChangesetId]
+  );
+
+  const updateVariationAtIndex = useCallback(
+    async (index: number, updates: Partial<VisualEditorVariation>) => {
+      const newVariation = {
+        ...variations[index],
+        ...updates,
+      };
+      const newVariations = [
+        ...(variations?.map((v, i) => (i === index ? newVariation : v)) ?? []),
+      ];
+
+      setVariations(newVariations);
+
+      updateVisualChangeset(newVariations);
+    },
+    [variations, setVariations, updateVisualChangeset]
+  );
+
+  // generate normalized variations
+  useEffect(() => {
+    if (!visualChangeset || !experiment) return;
+
+    const visualEditorVariations = normalizeVariations({
+      experiment,
+      visualChangeset,
+    });
+
+    setVariations(visualEditorVariations);
+  }, [visualChangeset, experiment]);
+
+  // listen for responses from background script
+  useEffect(() => {
+    if (!visualChangesetId) return;
+
+    const messageHandler = (event: MessageEvent<Message>) => {
+      const msg = event.data;
+      switch (msg.type) {
+        case "GB_RESPONSE_LOAD_VISUAL_CHANGESET":
+          if (msg.data.error) setError(msg.data.error);
+          else {
+            setVisualChangeset(msg.data.visualChangeset);
+            setExperiment(msg.data.experiment);
+            setExperimentUrl(msg.data.experimentUrl);
+          }
+          setLoading(false);
+          break;
+        case "GB_RESPONSE_UPDATE_VISUAL_CHANGESET":
+          if (msg.data.error) setError(msg.data.error);
+          else setVisualChangeset(msg.data.visualChangeset);
+          setLoading(false);
+          break;
+        default:
+          break;
+      }
+    };
+
+    window.addEventListener("message", messageHandler);
+
+    // load visual changeset on init
+    setLoading(true);
+
+    const loadVisualChangesetMessage: LoadVisualChangesetRequestMessage = {
+      type: "GB_REQUEST_LOAD_VISUAL_CHANGESET",
+      data: {
+        visualChangesetId,
+      },
+    };
+
+    window.postMessage(loadVisualChangesetMessage, window.location.origin);
+
+    return () => window.removeEventListener("message", messageHandler);
+  }, [visualChangesetId, setVisualChangeset, setExperiment]);
+
+  return {
+    loading,
+    experimentUrl,
+    variations,
+    updateVariationAtIndex,
+    error,
+    cspError,
+  };
+};
+
+export default useVisualChangeset;
